+++ VimR/VimR/AdvencedPref.swift
@@ -13,6 +13,7 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
     case setUseInteractiveZsh(Bool)
     case setUseSnapshotUpdate(Bool)
     case setNvimBinary(String)
+    case setNvimAppName(String)
   }
 
   let uuid = UUID()
@@ -31,6 +32,7 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
     self.useInteractiveZsh = state.mainWindowTemplate.useInteractiveZsh
     self.useSnapshotUpdate = state.useSnapshotUpdate
     self.nvimBinary = state.mainWindowTemplate.nvimBinary
+    self.nvimAppName = state.nvimAppName
 
     super.init(frame: .zero)
 
@@ -40,10 +42,12 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
     context.subscribe(uuid: self.uuid) { state in
       if self.useInteractiveZsh != state.mainWindowTemplate.useInteractiveZsh
         || self.nvimBinary != state.mainWindowTemplate.nvimBinary
+        || self.nvimAppName != state.nvimAppName
         || self.useSnapshotUpdate != state.useSnapshotUpdate
       {
         self.useInteractiveZsh = state.mainWindowTemplate.useInteractiveZsh
         self.nvimBinary = state.mainWindowTemplate.nvimBinary
+        self.nvimAppName = state.nvimAppName
         self.useSnapshotUpdate = state.useSnapshotUpdate
 
         self.updateViews()
@@ -56,10 +60,12 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
   private var useInteractiveZsh: Bool
   private var useSnapshotUpdate: Bool
   private var nvimBinary: String = ""
+  private var nvimAppName: String = ""
 
   private let useInteractiveZshCheckbox = NSButton(forAutoLayout: ())
   private let useSnapshotUpdateCheckbox = NSButton(forAutoLayout: ())
   private let nvimBinaryField = NSTextField(forAutoLayout: ())
+  private let nvimAppNameField = NSTextField(forAutoLayout: ())
 
   @available(*, unavailable)
   required init?(coder _: NSCoder) {
@@ -68,12 +74,14 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
 
   override func windowWillClose() {
     self.nvimBinaryFieldAction()
+    self.nvimAppNameFieldAction()
   }
 
   private func updateViews() {
     self.useSnapshotUpdateCheckbox.boolState = self.useSnapshotUpdate
     self.useInteractiveZshCheckbox.boolState = self.useInteractiveZsh
     self.nvimBinaryField.stringValue = self.nvimBinary
+    self.nvimAppNameField.stringValue = self.nvimAppName
   }
 
   private func addViews() {
@@ -105,8 +113,14 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
     of VimR in no time!
     """#)
 
-    let nvimBinaryTitle = self.titleTextField(title: "NeoVim Binary:")
+    let nvimBinaryTitle = self.titleTextField(title: "Nvim Binary:")
     let nvimBinaryField = self.nvimBinaryField
+
+    let nvimAppNameTitle = self.titleTextField(title: "NVIM_APPNAME:")
+    let nvimAppNameField = self.nvimAppNameField
+    let nvimAppNameInfo = self.infoTextField(markdown: #"""
+    When set, VimR will set the `NVIM_APPNAME` environment variable to this value by default.
+    """#)
 
     self.addSubview(paneTitle)
 
@@ -116,31 +130,34 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
     self.addSubview(useInteractiveZshInfo)
     self.addSubview(nvimBinaryTitle)
     self.addSubview(nvimBinaryField)
+    self.addSubview(nvimAppNameTitle)
+    self.addSubview(nvimAppNameField)
+    self.addSubview(nvimAppNameInfo)
 
     paneTitle.autoPinEdge(toSuperviewEdge: .top, withInset: 18)
     paneTitle.autoPinEdge(toSuperviewEdge: .left, withInset: 18)
     paneTitle.autoPinEdge(toSuperviewEdge: .right, withInset: 18, relation: .greaterThanOrEqual)
 
     useSnapshotUpdate.autoPinEdge(.top, to: .bottom, of: paneTitle, withOffset: 18)
-    useSnapshotUpdate.autoPinEdge(.left, to: .left, of: paneTitle)
+    useSnapshotUpdate.autoPinEdge(.left, to: .right, of: nvimAppNameTitle, withOffset: 5)
 
     useSnapshotUpdateInfo.autoPinEdge(.top, to: .bottom, of: useSnapshotUpdate, withOffset: 5)
     useSnapshotUpdateInfo.autoPinEdge(.left, to: .left, of: useSnapshotUpdate)
 
     useInteractiveZsh.autoPinEdge(.top, to: .bottom, of: useSnapshotUpdateInfo, withOffset: 18)
-    useInteractiveZsh.autoPinEdge(.left, to: .left, of: useSnapshotUpdate)
+    useInteractiveZsh.autoPinEdge(.left, to: .right, of: nvimAppNameTitle, withOffset: 5)
 
     useInteractiveZshInfo.autoPinEdge(.top, to: .bottom, of: useInteractiveZsh, withOffset: 5)
     useInteractiveZshInfo.autoPinEdge(.left, to: .left, of: useInteractiveZsh)
 
     nvimBinaryTitle.autoPinEdge(.top, to: .bottom, of: useInteractiveZshInfo, withOffset: 18)
-    nvimBinaryTitle.autoPinEdge(.left, to: .left, of: useSnapshotUpdate)
-    // nvimBinaryTitle.autoAlignAxis(.baseline, toSameAxisOf: nvimBinaryField)
-
-    nvimBinaryField.autoPinEdge(.top, to: .bottom, of: useInteractiveZshInfo, withOffset: 18)
-    nvimBinaryField.autoPinEdge(.left, to: .right, of: nvimBinaryTitle)
-    nvimBinaryField.autoPinEdge(toSuperviewEdge: .right, withInset: 18)
-    nvimBinaryField.autoSetDimension(.height, toSize: 20, relation: .greaterThanOrEqual)
+    nvimBinaryTitle.autoPinEdge(.right, to: .right, of: nvimAppNameTitle)
+    
+    nvimBinaryField.autoAlignAxis(.baseline, toSameAxisOf: nvimBinaryTitle)
+    nvimBinaryField.autoPinEdge(.left, to: .right, of: nvimBinaryTitle, withOffset: 5)
+    nvimBinaryField.autoSetDimension(.width, toSize: 180, relation: .greaterThanOrEqual)
+    nvimBinaryField.autoSetDimension(.width, toSize: 400, relation: .lessThanOrEqual)
+    nvimBinaryField.autoPinEdge(toSuperviewEdge: .right, withInset: 18, relation: .greaterThanOrEqual)
     NotificationCenter.default.addObserver(
       forName: NSControl.textDidEndEditingNotification,
       object: nvimBinaryField,
@@ -148,6 +165,23 @@ final class AdvancedPref: PrefPane, UiComponent, NSTextFieldDelegate {
     ) { [weak self] _ in
       Task { @MainActor in self?.nvimBinaryFieldAction() }
     }
+
+    nvimAppNameTitle.autoPinEdge(.top, to: .bottom, of: nvimBinaryField, withOffset: 18)
+    nvimAppNameTitle.autoPinEdge(toSuperviewEdge: .left, withInset: 18, relation: .greaterThanOrEqual)
+
+    nvimAppNameField.autoAlignAxis(.baseline, toSameAxisOf: nvimAppNameTitle)
+    nvimAppNameField.autoPinEdge(.left, to: .right, of: nvimAppNameTitle, withOffset: 5)
+    nvimAppNameField.autoSetDimension(.width, toSize: 180)
+    NotificationCenter.default.addObserver(
+      forName: NSControl.textDidEndEditingNotification,
+      object: nvimAppNameField,
+      queue: nil
+    ) { [weak self] _ in
+      Task { @MainActor in self?.nvimAppNameFieldAction() }
+    }
+
+    nvimAppNameInfo.autoPinEdge(.top, to: .bottom, of: nvimAppNameField, withOffset: 5)
+    nvimAppNameInfo.autoPinEdge(.left, to: .right, of: nvimAppNameTitle, withOffset: 5)
   }
 }
 
@@ -165,5 +199,10 @@ extension AdvancedPref {
   func nvimBinaryFieldAction() {
     let newNvimBinary = self.nvimBinaryField.stringValue
     self.emit(.setNvimBinary(newNvimBinary))
+  }
+
+  func nvimAppNameFieldAction() {
+    let newNvimAppName = self.nvimAppNameField.stringValue
+    self.emit(.setNvimAppName(newNvimAppName))
   }
 }